Address a problem with nc_put_att in debug mode (issue #11)
When using the debugging option (environment variable debug set to 1)
nc_put_att crashes because of an un-allocated variable. The reason for
this is that nc_put_att checks the size of v%dims before checking if v
is a coordinate or a data variable. The size of v%dims is only needed
for data variables, but the crash originates in calling nc_put_att for a
coordinate variable.

The solution is to check the size of v%dims only after making sure it's
a data variable, because v%dims is not allocated for coordinate
variables.